% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_frame.R
\name{clean_data_frame}
\alias{clean_data_frame}
\title{Clean Data Frame}
\usage{
clean_data_frame(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
Clean up data frame.
}
\description{
Clean Data Frame
}
