% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Merge taxa, water quality, and tab.habitat data into one}
\usage{
prep_data(tab.taxa, tab.wq, tab.habitat, tab.event, tab.stations, tab.project,
  tab.huc, agg_sample_num = TRUE, development = TRUE)
}
\arguments{
\item{tab.taxa}{= Taxonomic counts in long data format}

\item{tab.wq}{= Water quality data in a long data format}

\item{tab.habitat}{= tab.habitat data in a long data format}

\item{tab.event}{= EVENT_ID}

\item{tab.stations}{= Station information (STATION_ID, ECOREGION, STRAHLER)}

\item{agg_sample_num}{= Should the sample number be treated as seperate
samples (FALSE) or aggregated into a single sample (TRUE)?}

\item{development}{= during index development, "development" should be
set to TRUE.  When TRUE the sample_number (replicate) with the largest reporting
value (# of individual observed) is selected for.  During final scoring development
should be set to FALSE.  When FALSE all replicates are included in the final
output.}

\item{Taxonomy}{= TSN #'s with necessary taxonomic levels (CLASS, ORDER, FAMILY)}
}
\value{
Merge taxa water quality, and tab.habitat data into one large data frame
}
\description{
Merge taxa, water quality, and tab.habitat data into one
}
