% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ecoregion.R
\name{prep_ecoregion}
\alias{prep_ecoregion}
\title{Ecoregion and Bioregion Modifications}
\usage{
prep_ecoregion(long.df)
}
\arguments{
\item{long.df}{= Long data frame.}
}
\value{
Assigns Chessie BIBI related ecoregion and bioregion distinctions.
}
\description{
Ecoregion and Bioregion Modifications
}
