% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_habitat.R
\name{prep_habitat}
\alias{prep_habitat}
\title{Prepare Habitat Data}
\usage{
prep_habitat(long.df, agg_sample_num = TRUE)
}
\arguments{
\item{agg_sample_num}{= Should the sample number be treated as seperate
samples (FALSE) or aggregated into a single sample (TRUE)?}

\item{Habitat}{= Habitat data in long format}
}
\value{
Transform habitat data from long to wide format
}
\description{
Prepare Habitat Data
}
