% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_merge.R
\name{prep_merge}
\alias{prep_merge}
\title{Merge taxa, water quality, and habitat data into one}
\usage{
prep_merge(tab.event, taxa.prep, tab.stations, tab.wq, tab.habitat, tab.project,
  tab.huc)
}
\arguments{
\item{tab.event}{= EVENT_ID}

\item{taxa.prep}{= Taxa data in a long data format}

\item{tab.stations}{= Station information (STATION_ID, ECOREGION, STRAHLER)}

\item{tab.wq}{= Water quality data in a wide format}

\item{tab.habitat}{= Habitat data in a wide format}

\item{tab.project}{= Project information}

\item{tab.huc}{= HUC information merged on HUC-12}
}
\value{
Merge taxa water quality, and habitat data into one large data frame.
}
\description{
Merge taxa, water quality, and habitat data into one
}
