% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_method.R
\name{prep_method}
\alias{prep_method}
\title{Prepare subset specific to BIBI}
\usage{
prep_method(long.df)
}
\arguments{
\item{long.df}{Data in a long format.}
}
\value{
A subset of the data containing sampling events with strahler stream
order less than or equal to 4 and methods comparable to kick net samples.
}
\description{
Prepare subset specific to BIBI
}
