% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_taxa.R
\name{prep_taxa}
\alias{prep_taxa}
\title{Prepare Taxonomic Counts}
\usage{
prep_taxa(long.df)
}
\arguments{
\item{long.df}{Taxonomic counts in long data format.}
}
\value{
A data frame.
}
\description{
Prepare Taxonomic Counts
}
