% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_wq.R
\name{prep_wq}
\alias{prep_wq}
\title{Prepare Water Quality Data}
\usage{
prep_wq(long.df)
}
\arguments{
\item{long.df}{= Water quality data in long format}
}
\value{
Transform water quality data from long to wide format
}
\description{
Prepare Water Quality Data
}
