% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_classifiation.R
\name{site_classification}
\alias{site_classification}
\title{Prepare Class Multi}
\usage{
site_classification(long.df)
}
\arguments{
\item{long.df}{= the output of the prep_env function.}
}
\value{
A data frame containing environmental data to classify sites.
}
\description{
Prepare Class Multi
}
